Let --scale take a fraction; below 1 renders smaller than the cells
scale and want_scale become doubles, parsed with atof and clamped to 0.1..3.
The wide-pane cap only lowers a ratio above 1, the screencast size rounds to
whole pixels with a floor of one, and the s key still steps whole ratios.

+++ src/web.h
@@ -268,8 +268,8 @@ typedef struct {
 
     int     css_w, css_h;      // page viewport in CSS pixels
     int     cast_w, cast_h;    // pixel size the screencast was last asked for
-    int     scale;             // device pixel ratio actually in use
-    int     want_scale;        // what was asked for on the command line
+    double  scale;             // device pixel ratio actually in use
+    double  want_scale;        // what was asked for on the command line
     int     img_rows;          // cell rows the page occupies
 
     char    url[1024];
